start moving stats calculation into dedicated module
